Added user authentication
- User can login/logout
- User must be logged in before it can destroy/edit a record

+++ laravel/app/views/category/index.blade.php
@@ -11,7 +11,9 @@
         <tr>
             <th>ID</th>
             <th>Category</th>
-            <th>Actions</th>
+            @if(Auth::check())
+                <th>Actions</th>
+            @endif
         </tr>
         </thead>
         <tbody>
@@ -19,18 +21,20 @@
             <tr>
                 <td>{{ $category->id }}</td>
                 <td>{{ HTML::link( route('category.show', $category->id), $category->name) }}</td>
-                <td class="actions">
-                    {{ Form::open(['route' => ['category.edit', $category->id], 'method' => 'GET']) }}
-                    {{ Form::submit('Edit', ['class' => 'inline button tiny']) }}
-                    {{ Form::close() }}
-
-                    @if ( $category->id != 1 )
-                        {{ Form::open(['route' => ['category.destroy', $category->id], 'method' => 'DELETE']) }}
-                        {{ Form::token() }}
-                        {{ Form::submit('Destroy', ['class' => 'inline button tiny alert']) }}
+                @if(Auth::check())
+                    <td class="actions">
+                        {{ Form::open(['route' => ['category.edit', $category->id], 'method' => 'GET']) }}
+                        {{ Form::submit('Edit', ['class' => 'inline button tiny']) }}
                         {{ Form::close() }}
-                    @endif
-                </td>
+
+                        @if ( $category->id != 1 )
+                            {{ Form::open(['route' => ['category.destroy', $category->id], 'method' => 'DELETE']) }}
+                            {{ Form::token() }}
+                            {{ Form::submit('Destroy', ['class' => 'inline button tiny alert']) }}
+                            {{ Form::close() }}
+                        @endif
+                    </td>
+                @endif
             </tr>
         @endforeach
         </tbody>
